requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
attrs==22.1.0
    # via -r requirements.in
cachecontrol==0.12.11
    # via firebase-admin
cachetools==5.2.0
    # via google-auth
certifi==2022.9.24
    # via requests
cffi==1.15.1
    # via cryptography
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via flask
cryptography==38.0.3
    # via pyjwt
distlib==0.3.6
    # via virtualenv
filelock==3.8.0
    # via virtualenv
firebase-admin==6.0.1
    # via -r requirements.in
flask==2.2.2
    # via
    #   -r requirements.in
    #   flask-login
flask-login==0.6.2
    # via -r requirements.in
google-api-core[grpc]==2.10.2
    # via
    #   firebase-admin
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-datastore
    #   google-cloud-firestore
    #   google-cloud-storage
google-api-python-client==2.65.0
    # via firebase-admin
google-auth==2.14.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-cloud-core==2.3.2
    # via
    #   -r requirements.in
    #   google-cloud-datastore
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-datastore==1.15.5
    # via -r requirements.in
google-cloud-firestore==2.7.2
    # via firebase-admin
google-cloud-storage==2.5.0
    # via firebase-admin
google-crc32c==1.5.0
    # via google-resumable-media
google-resumable-media==2.4.0
    # via google-cloud-storage
googleapis-common-protos==1.56.4
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.50.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.48.2
    # via google-api-core
httplib2==0.21.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
identify==2.5.6
    # via pre-commit
idna==3.4
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
msgpack==1.0.4
    # via cachecontrol
nodeenv==1.7.0
    # via pre-commit
platformdirs==2.5.2
    # via virtualenv
pre-commit==2.20.0
    # via -r requirements.in
proto-plus==1.22.1
    # via google-cloud-firestore
protobuf==3.20.3
    # via
    #   google-api-core
    #   google-cloud-datastore
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.6.0
    # via firebase-admin
pyparsing==3.0.9
    # via httplib2
pyyaml==6.0
    # via pre-commit
requests==2.28.1
    # via
    #   cachecontrol
    #   google-api-core
    #   google-cloud-storage
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
    #   grpcio
toml==0.10.2
    # via pre-commit
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.12
    # via requests
virtualenv==20.16.5
    # via pre-commit
werkzeug==2.2.2
    # via
    #   flask
    #   flask-login

# The following packages are considered to be unsafe in a requirements file:
# setuptools
